2010-04-19  Yury Semikhatsky  <yurys@chromium.org>
        Unreviewed. Chromium build fix.

        * bindings/v8/JavaScriptCallFrame.h:


Canonical link: https://commits.webkit.org/49094@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@57813 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/bindings/v8/JavaScriptCallFrame.h
@@ -49,7 +49,7 @@ public:
 
     JavaScriptCallFrame* caller();
 
-    intptr_t sourceID() const;
+    int sourceID() const;
     int line() const;
     String functionName() const;
